Add a web view
Web view is supposed to allow people to take quick a glance at current
state and processed patches. It's not a main part of this app.

I've decided take a static approach for this web view. Templating pairs
well with SQLModel and for this very simple app and data it should be
enough.

I've also decided to use pico.css because it's plug and play solution
which requires zero to none css modifications, and it looks really
great.